Move lint utils to own package
