+++ pkg/lint/utils/testdata/external_ref.json
