+++ pkg/lint/utils/testdata/self_referencing_ref.json
