+++ pkg/lint/utils/testdata/with_external_ref.json
