+++ pkg/lint/utils/testdata/with_internal_defs_ref.json
